added authentication to users

+++ backend/controllers/todoController.js
@@ -1,12 +1,13 @@
 const asyncHandler = require("express-async-handler");
 
 const Todo = require("../models/TodoModel");
+const User = require("../models/UserModel");
 
 // @desc Get todos
 // @route GET /api/todos
 // @access Private
 const getTodos = asyncHandler(async (req, res) => {
-  const todos = await Todo.find();
+  const todos = await Todo.find({ user: req.user.id });
   res.status(200).json(todos);
 });
 
@@ -21,6 +22,7 @@ const setTodo = asyncHandler(async (req, res) => {
 
   const todo = await Todo.create({
     text: req.body.text,
+    user: req.user.id,
   });
 
   res.status(200).json(todo);
@@ -37,6 +39,20 @@ const updateTodo = asyncHandler(async (req, res) => {
     throw new Error("Todo not found!");
   }
 
+  const user = await User.findById(req.user.id);
+
+  //Check for user
+  if (!user) {
+    res.status(401);
+    throw new Error("User not found");
+  }
+
+  //Make sure the logged in user matches the todo user
+  if (todo.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error("User not authorized");
+  }
+
   const updatedTodo = await Todo.findByIdAndUpdate(req.params.id, req.body, {
     new: true,
   });
@@ -49,6 +65,25 @@ const updateTodo = asyncHandler(async (req, res) => {
 // @access Private
 const deleteTodo = asyncHandler(async (req, res) => {
   const todo = await Todo.findById(req.params.id);
+
+  if (!todo) {
+    res.status(400);
+    throw new Error("Todo not found");
+  }
+
+  const user = await User.findById(req.user.id);
+
+  //Check for user
+  if (!user) {
+    res.status(401);
+    throw new Error("User not found");
+  }
+
+  //Make sure the logged in user matches the todo user
+  if (todo.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error("User not authorized");
+  }
 
   await todo.deleteOne();
